Activity Melodies xml and activity
Hey there...My name is Margret Mauno. I am trying to develop an app that will make an impact on the lives of musicians and music lovers. All that I need to complete the app is to ...1....enable the melodies that display to actually play when each list item is clicked. I was also hoping to ......2....enable sharing of audio files. With the current code, the error ...file format not supported..... displays.... yet the file format is mp3 . Kindly help me solve these two errors and impact on the life of an instrumentalist and musician. I truly appreciate. Thank you  so much in advance- Kind regards, Margret.